// src/app/shared/interfaces/nuvei-query-params-response.interface.ts
export interface NuveiQueryParamsResponse {
    payment_status:          string;
    ppp_status:              string;
    cardCompany:             string;
    nameOnCard:              string;
    country:                 string;
    email:                   string;
    currency:                string;
    merchant_site_id:        string;
    merchant_id:             string;
    merchantLocale:          string;
    requestVersion:          string;
    PPP_TransactionID:       string;
    productId:               string;
    customData:              string;
    payment_method:          string;
    responseTimeStamp:       string;
    message:                 string;
    Error:                   string;
    "3Dflow":                string;
    userPaymentOptionId:     string;
    Status:                  string;
    ExErrCode:               string;
    ErrCode:                 string;
    AuthCode:                string;
    ReasonCode:              string;
    Token:                   string;
    tokenId:                 string;
    responsechecksum:        string;
    advanceResponseChecksum: string;
    totalAmount:             number;
    TransactionID:           string;
    dynamicDescriptor:       string;
    uniqueCC:                string;
    eci:                     string;
    orderTransactionId:      string;
    item_amount_1:           string;
    item_quantity_1:         string;
}